Move FHIR resources to folders
