+++ packages/shared-src/src/models/fhir/ImagingStudy/index.js
@@ -2,12 +2,16 @@ import config from 'config';
 import { DataTypes } from 'sequelize';
 import * as yup from 'yup';
 
-import { FhirResource } from './Resource';
+import { FhirResource } from '../Resource';
 
-import { FhirAnnotation, FhirIdentifier, FhirReference } from '../../services/fhirTypes';
-import { FHIR_INTERACTIONS, FHIR_ISSUE_TYPE, IMAGING_REQUEST_STATUS_TYPES } from '../../constants';
-import { Deleted, Invalid } from '../../utils/fhir';
-import { getCurrentDateTimeString, toDateTimeString } from '../../utils/dateTime';
+import { FhirAnnotation, FhirIdentifier, FhirReference } from '../../../services/fhirTypes';
+import {
+  FHIR_INTERACTIONS,
+  FHIR_ISSUE_TYPE,
+  IMAGING_REQUEST_STATUS_TYPES,
+} from '../../../constants';
+import { Deleted, Invalid } from '../../../utils/fhir';
+import { getCurrentDateTimeString, toDateTimeString } from '../../../utils/dateTime';
 
 export class FhirImagingStudy extends FhirResource {
   static init(options, models) {
