+++ packages/shared-src/src/models/fhir/JobWorker/index.js
@@ -1,6 +1,6 @@
 import Sequelize, { DataTypes, QueryTypes } from 'sequelize';
-import { Model } from '../Model';
-import { SYNC_DIRECTIONS } from '../../constants';
+import { Model } from '../../Model';
+import { SYNC_DIRECTIONS } from '../../../constants';
 
 export class FhirJobWorker extends Model {
   static init({ primaryKey, ...options }) {
